+++ test/typechecking/unsupported/type_extension_with_spec.mli
@@ -1,0 +1,8 @@
+type t = ..
+(*@ model : integer *)
+(* {gospel_expected|
+[125] gospel: internal error, uncaught exception:
+              File "src/uattr2spec.ml", line 102, characters 9-15: Assertion failed
+              
+      
+|gospel_expected} *)